dm/requirements.txt
Django==2.0.7
PyMySQL==0.9.3
djangorestframework==3.9.1
django-rest-auth
django-background-tasks
django-crum
django-filter